+++ hypervisor/arch/x86/notify.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2022 Intel Corporation.
+ * Copyright (C) 2018-2025 Intel Corporation.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -14,70 +14,9 @@
 #include <asm/lapic.h>
 #include <asm/guest/vm.h>
 #include <asm/guest/virq.h>
+#include <common/notify.h>
 
 static uint32_t notification_irq = IRQ_INVALID;
-
-static uint64_t smp_call_mask = 0UL;
-
-/* run in interrupt context */
-static void kick_notification(__unused uint32_t irq, __unused void *data)
-{
-	/* Notification vector is used to kick target cpu out of non-root mode.
-	 * And it also serves for smp call.
-	 */
-	uint16_t pcpu_id = get_pcpu_id();
-
-	if (bitmap_test(pcpu_id, &smp_call_mask)) {
-		struct smp_call_info_data *smp_call =
-			&per_cpu(smp_call_info, pcpu_id);
-
-		if (smp_call->func != NULL) {
-			smp_call->func(smp_call->data);
-		}
-		bitmap_clear_lock(pcpu_id, &smp_call_mask);
-	}
-}
-
-void handle_smp_call(void)
-{
-	kick_notification(0, NULL);
-}
-
-void smp_call_function(uint64_t mask, smp_call_func_t func, void *data)
-{
-	uint16_t pcpu_id;
-	struct smp_call_info_data *smp_call;
-
-	/* wait for previous smp call complete, which may run on other cpus */
-	while (atomic_cmpxchg64(&smp_call_mask, 0UL, mask) != 0UL);
-	pcpu_id = ffs64(mask);
-	while (pcpu_id < MAX_PCPU_NUM) {
-		bitmap_clear_nolock(pcpu_id, &mask);
-		if (pcpu_id == get_pcpu_id()) {
-			func(data);
-			bitmap_clear_nolock(pcpu_id, &smp_call_mask);
-		} else if (is_pcpu_active(pcpu_id)) {
-			smp_call = &per_cpu(smp_call_info, pcpu_id);
-			smp_call->func = func;
-			smp_call->data = data;
-
-			struct acrn_vcpu *vcpu = get_ever_run_vcpu(pcpu_id);
-
-			if ((vcpu != NULL) && (is_lapic_pt_enabled(vcpu))) {
-				vcpu_make_request(vcpu, ACRN_REQUEST_SMP_CALL);
-			} else {
-				arch_send_single_ipi(pcpu_id, NOTIFY_VCPU_VECTOR);
-			}
-		} else {
-			/* pcpu is not in active, print error */
-			pr_err("pcpu_id %d not in active!", pcpu_id);
-			bitmap_clear_nolock(pcpu_id, &smp_call_mask);
-		}
-		pcpu_id = ffs64(mask);
-	}
-	/* wait for current smp call complete */
-	wait_sync_change(&smp_call_mask, 0UL);
-}
 
 static int32_t request_notification_irq(irq_action_t func, void *data)
 {
@@ -103,7 +42,7 @@ static int32_t request_notification_irq(irq_action_t func, void *data)
 /*
  * @pre be called only by BSP initialization process
  */
-void setup_notification(void)
+static void setup_notification(void)
 {
 	/* support IPI notification, Service VM will register all CPU */
 	if (request_notification_irq(kick_notification, NULL) < 0) {
@@ -127,7 +66,7 @@ static void handle_pi_notification(uint32_t irq, __unused void *data)
 }
 
 /*pre-condition: be called only by BSP initialization proccess*/
-void setup_pi_notification(void)
+static void setup_pi_notification(void)
 {
 	uint32_t i;
 
@@ -136,5 +75,22 @@ void setup_pi_notification(void)
 			pr_err("Failed to setup pi notification");
 			break;
 		}
+	}
+}
+
+void arch_init_smp_call(void)
+{
+	setup_notification();
+	setup_pi_notification();
+}
+
+void arch_smp_call_kick_pcpu(uint16_t pcpu_id)
+{
+	struct acrn_vcpu *vcpu = get_ever_run_vcpu(pcpu_id);
+
+	if ((vcpu != NULL) && (is_lapic_pt_enabled(vcpu))) {
+		vcpu_make_request(vcpu, ACRN_REQUEST_SMP_CALL);
+	} else {
+		arch_send_single_ipi(pcpu_id, NOTIFY_VCPU_VECTOR);
 	}
 }
