+++ hypervisor/include/arch/x86/asm/notify.h
@@ -1,23 +1,16 @@
 /*
- * Copyright (C) 2020-2022 Intel Corporation.
+ * Copyright (C) 2020-2025 Intel Corporation.
  *
  * SPDX-License-Identifier: BSD-3-Clause
+ *
  */
 
-#ifndef NOTIFY_H
-#define NOTIFY_H
+#ifndef X86_NOTIFY_H
+#define X86_NOTIFY_H
 
-typedef void (*smp_call_func_t)(void *data);
-struct smp_call_info_data {
-	smp_call_func_t func;
-	void *data;
-};
+#include <types.h>
 
-struct acrn_vm;
-void smp_call_function(uint64_t mask, smp_call_func_t func, void *data);
+void arch_init_smp_call(void);
+void arch_smp_call_kick_pcpu(uint16_t pcpu_id);
 
-void setup_notification(void);
-void handle_smp_call(void);
-void setup_pi_notification(void);
-
-#endif
+#endif /* X86_NOTIFY_H */
